03_code/requirements.txt
Pillow>=10.0.0
numpy>=1.26.0
matplotlib>=3.8.0
scikit-learn>=1.4.0
PyYAML>=6.0.0
